+++ lib/ChargeBee/Result.php
@@ -32,7 +32,7 @@ class ChargeBee_Result
     function invoice() 
     {
         return $this->_get('invoice', 'ChargeBee_Invoice', 
-        array('line_items' => 'ChargeBee_InvoiceLineItem', 'discounts' => 'ChargeBee_InvoiceDiscount', 'taxes' => 'ChargeBee_InvoiceTax', 'invoice_transactions' => 'ChargeBee_InvoiceLinkedTransaction', 'orders' => 'ChargeBee_InvoiceLinkedOrder'));
+        array('line_items' => 'ChargeBee_InvoiceLineItem', 'discounts' => 'ChargeBee_InvoiceDiscount', 'taxes' => 'ChargeBee_InvoiceTax', 'invoice_transactions' => 'ChargeBee_InvoiceLinkedTransaction', 'orders' => 'ChargeBee_InvoiceLinkedOrder', 'shipping_address' => 'ChargeBee_InvoiceShippingAddress', 'billing_address' => 'ChargeBee_InvoiceBillingAddress'));
     }
 
     function order() 
@@ -99,7 +99,8 @@ class ChargeBee_Result
 
     function portalSession() 
     {
-        return $this->_get('portal_session', 'ChargeBee_PortalSession');
+        return $this->_get('portal_session', 'ChargeBee_PortalSession', 
+        array('linked_customers' => 'ChargeBee_PortalSessionLinkedCustomer'));
     }